Update
The class name has changed to SetAccount and a new class named GetAccount has been added. With the SetAccount class, we can create a password and save it to the database in an encrypted form. With the GetAccount class, we can retrieve the encrypted password and account information from the database.